dtype changes made

+++ testsuite/MDAnalysisTests/lib/test_cutil.py
@@ -50,7 +50,7 @@ def test_unique_int_1d(values):
 
 def test_array_unique_int_1d():
     try:
-        res_indices = np.array([6, 5, 5, 7], dtype=np.int32)
+        res_indices = np.array([6, 5, 5, 7], dtype=np.int_)
         assert_equal(unique_int_1d(res_indices),  [5, 6, 7])
     except ValueError:
         raise AssertionError("Issue #2687 failure")
